requirements.txt
groq
openai
pydub
pyttsx3
gspread
pyaudio
SpeechRecognition
oauth2client
python-dotenv
agno
duckduckgo-search